chore(hooks): extract fee calculation into seperate store
