+++ src/store/transactionStore.ts
@@ -1,4 +1,4 @@
-import { EthArgs, UnmarshalledFees } from "@renproject/interfaces";
+import { EthArgs } from "@renproject/interfaces";
 import RenJS from "@renproject/ren";
 import { AbiItem } from "web3-utils";
 import { LockAndMint } from "@renproject/ren/build/main/lockAndMint";
@@ -9,10 +9,9 @@ import { useCallback, useEffect, useState } from "react";
 
 import { Transaction } from "../types/transaction";
 import adapterABI from "../utils/ABIs/adapterCurveABI.json";
-import curveABI from "../utils/ABIs/curveABI.json";
 import { Asset } from "../utils/assets";
-import { CURVE_MAIN, CURVE_TEST } from "../utils/environmentVariables";
 import { Store } from "./store";
+import { FeeStore } from "./feeStore";
 
 function useTransactionStore() {
   const {
@@ -20,19 +19,14 @@ function useTransactionStore() {
     fsEnabled,
     localWeb3Address,
     fsSignature,
-    dataWeb3,
     selectedNetwork,
-    fees,
     localWeb3,
     sdk,
 
     convertTransactions,
-    convertAmount,
-    convertSelectedDirection,
     convertPendingConvertToEthereum,
     convertAdapterAddress,
 
-    setFees,
     setSwapRevertModalTx,
     setSwapRevertModalExchangeRate,
     setShowSwapRevertModal,
@@ -40,12 +34,12 @@ function useTransactionStore() {
     setGatewayModalTx,
 
     setConvertTransactions,
-    setConvertExchangeRate,
-    setConvertRenVMFee,
-    setConvertNetworkFee,
-    setConvertConversionTotal,
     setConvertPendingConvertToEthereum,
   } = Store.useContainer();
+
+  const {
+    getFinalDepositExchangeRate,
+  } = FeeStore.useContainer();
 
   // Changing TX State
   const addTx = useCallback(
@@ -131,115 +125,6 @@ function useTransactionStore() {
     return convertTransactions.filter((t) => t.id === tx.id).size > 0;
   };
 
-  // External Data
-  const updateRenVMFees = async () => {
-    try {
-      const fees = await fetch("https://lightnode-mainnet.herokuapp.com", {
-        method: "POST", // or 'PUT'
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          id: 67,
-          jsonrpc: "2.0",
-          method: "ren_queryFees",
-          params: {},
-        }),
-      });
-      const data: UnmarshalledFees = (await fees.json()).result;
-      setFees(data);
-    } catch (e) {
-      console.error(e);
-    }
-  };
-
-  const getFinalDepositExchangeRate = useCallback(
-    async (tx: Transaction) => {
-      const { renResponse } = tx;
-
-      const utxoAmountInSats = Number(renResponse.autogen.amount);
-      const dynamicFeeRate = Number(fees![Asset.BTC].ethereum["mint"] / 10000);
-      const finalAmount = Math.round(utxoAmountInSats * (1 - dynamicFeeRate));
-
-      const curve = new dataWeb3!.eth.Contract(
-        curveABI as AbiItem[],
-        selectedNetwork === "testnet" ? CURVE_TEST : CURVE_MAIN
-      );
-      try {
-        const swapResult = await curve.methods.get_dy(0, 1, finalAmount).call();
-        return Number(swapResult / finalAmount);
-      } catch (e) {
-        console.error(e);
-      }
-    },
-    [dataWeb3, fees, selectedNetwork]
-  );
-
-  const gatherFeeData = async () => {
-    const amount = convertAmount;
-    const selectedDirection = convertSelectedDirection;
-    const fixedFeeKey = selectedDirection ? "release" : "lock";
-    const dynamicFeeKey = selectedDirection ? "burn" : "mint";
-
-    const fixedFee = Number(fees![Asset.BTC][fixedFeeKey] / 10 ** 8);
-    const dynamicFeeRate = Number(
-      fees![Asset.BTC].ethereum[dynamicFeeKey] / 10000
-    );
-
-    if (!amount || !dataWeb3 || !fees) return;
-
-    try {
-      let exchangeRate: number;
-      let renVMFee: number;
-      let total: number | string;
-      const amountInSats = Math.round(
-        RenJS.utils.value(amount, Asset.BTC).sats().toNumber()
-      );
-      const curve = new dataWeb3.eth.Contract(
-        curveABI as AbiItem[],
-        selectedNetwork === "testnet" ? CURVE_TEST : CURVE_MAIN
-      );
-
-      // withdraw
-      if (selectedDirection) {
-        const swapResult =
-          (await curve.methods.get_dy(1, 0, amountInSats).call()) / 10 ** 8;
-        exchangeRate = Number(swapResult / Number(amount));
-        renVMFee = Number(swapResult) * dynamicFeeRate;
-        total =
-          Number(swapResult - renVMFee - fixedFee) > 0
-            ? Number(swapResult - renVMFee - fixedFee)
-            : "0.000000";
-      } else {
-        renVMFee = Number(amount) * dynamicFeeRate;
-        const amountAfterMint =
-          Number(Number(amount) - renVMFee - fixedFee) > 0
-            ? Number(Number(amount) - renVMFee - fixedFee)
-            : 0;
-        const amountAfterMintInSats = Math.round(
-          RenJS.utils.value(amountAfterMint, Asset.BTC).sats().toNumber()
-        );
-
-        if (amountAfterMintInSats) {
-          const swapResult =
-            (await curve.methods.get_dy(0, 1, amountAfterMintInSats).call()) /
-            10 ** 8;
-          exchangeRate = Number(swapResult / amountAfterMint);
-          total = Number(swapResult);
-        } else {
-          exchangeRate = Number(0);
-          total = Number(0);
-        }
-      }
-
-      setConvertExchangeRate(exchangeRate);
-      setConvertRenVMFee(renVMFee);
-      setConvertNetworkFee(fixedFee);
-      setConvertConversionTotal(total);
-    } catch (e) {
-      console.error(e);
-    }
-  };
 
   // BTC to WBTC
   const monitorMintTx = useCallback(
@@ -729,8 +614,6 @@ function useTransactionStore() {
   return {
     updateTx,
     removeTx,
-    updateRenVMFees,
-    gatherFeeData,
     completeConvertToEthereum,
     initConvertToEthereum,
     initConvertFromEthereum,
